fix: внесены исправления в reducers
