+++ src/services/actions/actions.js
@@ -27,7 +27,7 @@ export function getIngredientsBurger() {
       .then((res) => {
         dispatch({
           type: GET_INGREDIENTS_SUCCESS,
-          ingredients: res.data
+          data: res.data
         })
       })
       .catch(() => {
