+++ src/services/reducers/ingredient-details.js
@@ -1,0 +1,31 @@
+import {
+  ADD_INGREDIENT_DATA,
+  DELETE_INGREDIENT_DATA,
+} from '../actions/actions';
+
+const initialState = {
+  modal: false,
+  ingredient: {},
+};
+
+export const ingredientReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ADD_INGREDIENT_DATA: {
+      return {
+        ...state,
+        modal: true,
+        ingredient: action.ingredient
+      }
+    }
+    case DELETE_INGREDIENT_DATA: {
+      return {
+        ...state,
+        modal: false,
+        ingredient: {}
+      }
+    }
+    default: {
+      return state;
+    }
+  }
+}
